fix #81: Modify Extension folder's naming rule
