+++ Front/Daily/Extension/PresentationDetent+extension.swift
@@ -1,0 +1,13 @@
+//
+//  PresentationDetent+extension.swift
+//  Daily
+//
+//  Created by seungyooooong on 10/20/24.
+//
+
+import Foundation
+import SwiftUI
+
+extension PresentationDetent {
+    static let small = Self.height(100)
+}
