+++ Front/Daily/Extension/UNUserNotificationCenter+extension.swift
@@ -1,8 +1,8 @@
 //
-//  UNUserNotificationCenter.swift
+//  UNUserNotificationCenter+extension.swift
 //  Daily
 //
-//  Created by 최승용 on 5/29/24.
+//  Created by seungyooooong on 10/20/24.
 //
 
 import Foundation
@@ -10,15 +10,11 @@ import UserNotifications
 
 extension UNUserNotificationCenter {
     func addNotiRequest(by date: DateComponents, id: String, title: String, body: String) {
-        
         let content = UNMutableNotificationContent()
         content.title = title
         content.body = body
         content.sound = .default
-//        content.badge = 1
-        
         let trigger = UNCalendarNotificationTrigger(dateMatching: date, repeats: true)
-        
         let request = UNNotificationRequest(identifier: id, content: content, trigger: trigger)
         
         self.add(request) { error in
