+++ Front/Daily/Extension/View+extension.swift
@@ -1,74 +1,41 @@
 //
-//  FrameExtension.swift
+//  View+extension.swift
 //  Daily
 //
-//  Created by 최승용 on 3/22/24.
+//  Created by seungyooooong on 10/20/24.
 //
 
 import Foundation
 import SwiftUI
 
-// MARK: - ResponsiveScreen
-
-extension CGFloat {
-    static let screenWidth = UIScreen.main.bounds.width
-    static let screenHeight = UIScreen.main.bounds.height
-    
-    // calendar
-    static let monthOnYearWidth = screenWidth / 3
-    static let monthOnYearHeight = screenHeight / 6
-    static let dayOnMonthWidth = screenWidth / 7
-    
-    static let fontSize = 6 * screenWidth / 393 // 6.7 iPhone 기준
-}
-
 // MARK: - Frame Modifier
-
 extension View {
-    // Vertical Center
-    func vCenter() -> some View {
-        self.frame(maxHeight: .infinity, alignment: .center)
-    }
-    
-    // Vertical Top
     func vTop() -> some View {
         self.frame(maxHeight: .infinity, alignment: .top)
     }
-    
-    // Vertical Bottom
+    func vCenter() -> some View {
+        self.frame(maxHeight: .infinity, alignment: .center)
+    }
     func vBottom() -> some View {
         self.frame(maxHeight: .infinity, alignment: .bottom)
     }
-    
-    // Horizontal Center
-    func hCenter() -> some View {
-        self.frame(maxWidth: .infinity, alignment: .center)
-    }
-    
-    // Horizontal Leading
     func hLeading() -> some View {
         self.frame(maxWidth: .infinity, alignment: .leading)
     }
-    
-    // Horizontal Trailing
+    func hCenter() -> some View {
+        self.frame(maxWidth: .infinity, alignment: .center)
+    }
     func hTrailing() -> some View {
         self.frame(maxWidth: .infinity, alignment: .trailing)
     }
 }
 
-// MARK: - Custom Sheet Size
-
-extension PresentationDetent {
-    static let small = Self.height(100)
-}
-
 // MARK: - Custom Gesture
-
 extension View {
     func mainViewDragGesture(userInfoViewModel: UserInfoViewModel, calendarViewModel: CalendarViewModel, alertViewModel: AlertViewModel) -> some View {
         self.gesture(
             DragGesture().onEnded { value in
-                // 좌 -> 우
+                // MARK: swipe to right
                 if value.translation.width > CGFloat.fontSize * 15 {
                     if value.startLocation.x < CGFloat.fontSize * 5 && calendarViewModel.getCurrentState() != "year" {
                         if calendarViewModel.getCurrentState() == "month" {
@@ -87,7 +54,7 @@ extension View {
                         }
                     }
                 }
-                // 우 -> 좌
+                // MARK: swipe to left
                 if value.translation.width < -CGFloat.fontSize * 15 {
                     calendarViewModel.changeCalendar(amount: 1, userInfoViewModel: userInfoViewModel) { code in
                         if code == "99" { alertViewModel.showAlert() }
@@ -104,3 +71,4 @@ extension View {
         UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
     }
 }
+
